grandchild_be/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
filelock==3.12.0
gunicorn==20.1.0
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
platformdirs==3.5.0
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
uritemplate==4.1.1
virtualenv==20.23.0
